Move cardano-wallet-api deps out of ui:common
